The find command's 'Permission denied' is finally corrected
There was a small bug where the 'Permission denied' message was displayed when a folder where the user's has no read rights was found by the 'find' command (it happened also on my Ubuntu 20.04 distro, it's not specific to WSL).

I fixed this bug by modifiying the "ModuleInitializer_FindPath" function, and by calling it in the main module's "Project.conf" file (it took me a while to figure why the aforementioned function didn't worked).